move utils.ts to types.ts
